Reorganize buffer and renderer classes
