+++ packages/graphics_test/src/GlobalTestFixture.cpp
@@ -1,5 +1,5 @@
 #include <graphics_test/GlobalTestFixture.h>
-#include <graphics/core/Renderer.h>
+#include <graphics/core/renderer/Renderer.h>
 #include <graphics/core/windows/Window.h>
 
 RendererPtr GlobalTestFixture::RENDERER = nullptr;
