+++ packages/graphics_test/src/core/renderer/RendererTest.cpp
@@ -1,6 +1,6 @@
 #include <boost/test/unit_test.hpp>
 #include <graphics_test/GlobalTestFixture.h>
-#include <graphics/core/Renderer.h>
+#include <graphics/core/renderer/Renderer.h>
 #include <graphics/core/windows/Window.h>
 #include <graphics/scene/Scene.h>
 #include <graphics/cameras/PerspectiveCamera.h>
