+++ packages/graphics_test/src/core/windows/WindowTest.cpp
@@ -1,6 +1,6 @@
 #include <boost/test/unit_test.hpp>
 #include <graphics_test/GlobalTestFixture.h>
-#include <graphics/core/Renderer.h>
+#include <graphics/core/renderer/Renderer.h>
 #include <graphics/core/windows/Window.h>
 
 /**
